fix: fix app with curr schema
