+++ python/timsseek_rts_receiver/pyproject.toml
@@ -10,6 +10,7 @@ dependencies = [
     "matplotlib",
     "requests",
     "speclib_builder[ml]",
+    "pydantic>=2.0.0",
 ]
 
 [build-system]
