+++ rust/timsquery/src/models/indices/transposed_quad_index/quad_splitted_transposed_index.rs
@@ -141,7 +141,6 @@ impl QuadSplittedTransposedIndex {
             }
         }
         IncludedRange::new(min, max)
-
     }
 }
